requirements.txt
bottle==0.12.21
cassandra-driver
